Fix:Post의 좋아요, 댓글 type 수정

+++ back/DevC/src/main/java/com/back/devc/domain/post/post/dto/AdminPostListResponse.java
@@ -12,8 +12,8 @@ public record AdminPostListResponse(
         Long categoryId,
         Boolean isDeleted,
         int viewCount,
-        long likeCount,
-        long commentCount,
+        int likeCount,
+        int commentCount,
         LocalDateTime createdAt
 ) {
     public AdminPostListResponse(Post post) {
